Added create and edit portfolio function

+++ backend/prisma/migrations/0_init/migration.sql
@@ -11,7 +11,7 @@ CREATE TABLE "Portfolio" (
     "id" SERIAL NOT NULL,
     "user_id" INTEGER NOT NULL,
     "crypto_id" INTEGER NOT NULL,
-    "quantity" INTEGER NOT NULL DEFAULT 0,
+    "quantity" DECIMAL(27,18) NOT NULL DEFAULT 0,
     "created_at" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updated_at" TIMESTAMP(3),
